React migrate

* Transition to Spring.

* Transition to Spring.

* Slight changes.

* Slight changes.
    - Deleted needless property files
    - Configured DataSource for liquibase
    - Configured Environment for email
    - Added getters and setters for entities
    - Rewrote services with dtos
    - Added ModelMapper

* Slight changes.
    - Deleted needless CommonService file distributing his duties
    - Added getters and setters annotations for entities and dtos
    - Rewrote services with dtos

* Slight changes.
    - Deleted needless dependencies
    - Implemented messaging with JMS

* Slight changes.
    - Implemented messaging with spring-boot-starter-mail

* Slight changes.
    - Implemented messaging with spring-boot-starter-mail

* Migrate to React

* Migrate to React
    Implemented shared state between Add and Phone components.

* Migrate to React
    Implemented add, edit, delete phones in Phone component.

* Migrate to React
    Implemented add, edit, delete phones in Attachment component.

* Migrate to React
    Implemented separate UserData component.

* Migrate to React
    Implemented About, Edit components.

* Migrate to React
    Deleted needless

* Migrate to React
    Implemented Email component.

* Migrate to React
    Implemented Add, About functions in backend.

* Migrate to React
    Implemented all CRUD operations for UI and backend.

* Migrate to React
    Implemented Email sending.

* Migrate to React
    Implemented Search function.

* Migrate to React
    Implemented Photo adding.

* Migrate to React
    Fixed "component is changing from uncontrolled to controlled"

* Migrate to React
    Implemented upload photo and attachments.

* Migrate to React
    Implemented attachments download.

* Migrate to React
    Implemented delete attachments from filesystem and upload during edit.

* Fix.
    Fixed explicit id setting during update.

* Fix.
    Changed all Long types with primitive.

* Fix.
    Expanded import statements
    Used try with resources
    Logged exception
    Simplified if else block

* Proxy for react.
    Added proxy for react and removed needless localhost.

* Refactor.
    Model mapper refactored.

* Refactor.
    Rid of @jsonignoreproperties.
    Transferred phone, attachment add and update concern to person service.

Approved-by: Сергей Кулак <kulaksd@gmail.com>

+++ core/src/main/java/com/itechart/contacts/core/filemanager/FileManageService.java
@@ -20,5 +20,5 @@ public interface FileManageService {
 
     void deleteFiles(int personId, long attachId);
 
-    void deleteUsers(int[] persons);
+    void deleteUsers(long[] persons);
 }
